testing migraphx_tf py independently
