+++ src/include/migraphx/tf.hpp
@@ -7,25 +7,7 @@
 namespace migraphx {
 inline namespace MIGRAPHX_INLINE_NS {
 
-struct unknown
-{
-    std::string op;
-    std::string name() const { return "unknown:" + op; }
-    shape compute_shape(std::vector<shape> input) const
-    {
-        if(input.empty())
-            return {};
-        else
-            return input.front();
-    }
-    friend std::ostream& operator<<(std::ostream& os, const unknown& x)
-    {
-        os << x.name();
-        return os;
-    }
-};
-
-/// Create a program from an onnx file
+/// Create a program from a tf pb file (default is nhwc format)
 program parse_tf(const std::string& name, bool is_nhwc);
 
 } // namespace MIGRAPHX_INLINE_NS
